microservices/auth/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pydantic
passlib[bcrypt]
bcrypt
python-dotenv
PyJWT